feat(wechat): 将客户消息独立为子模块并支持令牌失效自动刷新
